Replace ssize_t with long int for MinGW

+++ src/tooltable.c
@@ -39,7 +39,7 @@ static int have_tools_file = 0;
 static double tools[1+MAX_TOOL_NUMBER];
 
 static void 
-ProcessToolLine(const char *cp, const char *file_name, ssize_t file_line)
+ProcessToolLine(const char *cp, const char *file_name, long int file_line)
 {
     const char *cp0 = cp;
     int toolNumber;
@@ -56,13 +56,13 @@ ProcessToolLine(const char *cp, const char *file_name, ssize_t file_line)
     
     if (*cp != 'T') {
         GERB_COMPILE_WARNING(_("Ignored strange tool \"%s\" "
-				"at line %zd in file \"%s\""),
+				"at line %ld in file \"%s\""),
 			cp0, file_line, file_name);
         return;
     }
     if ((!isdigit((int) cp[1])) || (!isdigit((int) cp[2]))) {
         GERB_COMPILE_WARNING(_("No tool number in \"%s\" "
-				"at line %zd in file \"%s\""),
+				"at line %ld in file \"%s\""),
 			cp0, file_line, file_name);
         return;
     }
@@ -74,7 +74,7 @@ ProcessToolLine(const char *cp, const char *file_name, ssize_t file_line)
         toolNumber = atoi(tnb);
         if ((toolNumber < MIN_TOOL_NUMBER) || (toolNumber > MAX_TOOL_NUMBER)) {
             GERB_COMPILE_WARNING(_("Can't parse tool number in \"%s\" "
-				    "at line %zd in file \"%s\""),
+				    "at line %ld in file \"%s\""),
 			    cp0, file_line, file_name);
             return;
         }
@@ -93,19 +93,19 @@ ProcessToolLine(const char *cp, const char *file_name, ssize_t file_line)
 
     if (toolDia <= 0) {
         GERB_COMPILE_ERROR(_("Tool T%02d diameter is impossible "
-				"at line %zd in file \"%s\""),
+				"at line %ld in file \"%s\""),
 			toolNumber, file_line, file_name);
         return;
     }
     if (toolDia < 0.001) {
         GERB_COMPILE_WARNING(_("Tool T%02d diameter is very small "
-				"at line %zd in file \"%s\""),
+				"at line %ld in file \"%s\""),
 			toolNumber, file_line, file_name);
     }
     
     if (tools[toolNumber] != 0) {
         GERB_COMPILE_ERROR(_("Tool T%02d is already defined, occurred "
-				"at line %zd in file \"%s\""),
+				"at line %ld in file \"%s\""),
 			toolNumber, file_line, file_name);
         GERB_FATAL_ERROR(_("Exiting because this is a HOLD error "
 				"at any board house."));
@@ -122,7 +122,7 @@ gerbv_process_tools_file(const char *tf)
 {
     FILE *f;
     char buf[80];
-    ssize_t file_line = 0;
+    long int file_line = 0;
     
     have_tools_file = 0;
     memset(tools, 0, sizeof(tools));
